misc: Switch to using machine.Pin instances in lieu of ints.
It's more idiomatic of MicroPython to pass around a machine.Pin object
since they are singleton, will compare equal and are much more readily
useful.

While pins in machine.Pin.board don't get initialised in a useful
state we can init them all in a C module (in this case powman does it)
so they are useful right from boot.

Additionally rename everything from SW_ to BUTTON_ because a switch is
a thing that latches between two states and I will not concede on this.